requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.1
pillow==10.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
